asset config for moonbase
